services/privacy_service/requirements.txt
fastapi[standard]>=0.116.1
pydantic>=2.11.7
presidio-analyzer>=2.2.359
presidio-anonymizer>=2.2.359
pip>=25.2
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-instrumentation-fastapi>=0.57b0
opentelemetry-exporter-gcp-trace>=1.9.0
google-cloud-storage>=3.3.0
google-cloud-firestore>=2.21.0